src/AppLogistics.Controllers/Configuration/Novelties/NoveltiesController.cs
using AppLogistics.Components.Mvc;
using AppLogistics.Objects;
using AppLogistics.Services;
using AppLogistics.Validators;
using Microsoft.AspNetCore.Mvc;

namespace AppLogistics.Controllers.Configuration
{
    [Area("Configuration")]
    public class NoveltiesController : ValidatedController<INoveltyValidator, INoveltyService>
    {
        public NoveltiesController(INoveltyValidator validator, INoveltyService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create([BindExcludeId] NoveltyView novelty)
        {
            if (!Validator.CanCreate(novelty))
            {
                return View(novelty);
            }

            Service.Create(novelty);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(int id)
        {
            return NotEmptyView(Service.Get<NoveltyView>(id));
        }

        [HttpGet]
        public ActionResult Edit(int id)
        {
            return NotEmptyView(Service.Get<NoveltyView>(id));
        }

        [HttpPost]
        public ActionResult Edit(NoveltyView novelty)
        {
            if (!Validator.CanEdit(novelty))
            {
                return View(novelty);
            }

            Service.Edit(novelty);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(int id)
        {
            return NotEmptyView(Service.Get<NoveltyView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(int id)
        {
            if (!Validator.CanDelete(id))
            {
                return RedirectToAction("Delete", new { id });
            }

            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
